move cli parts index.js > cli.js
